requirements.txt
torch
numpy
torchvision
huggingface_hub
pillow
mediapipe
opencv-python

